requirement.txt
python-dotenv
pandas
numpy
pymongo[srv]==3.11
pymongo
certifi

## -e .